+++ src/user/user.controller.ts
@@ -7,6 +7,7 @@ import {
   Param,
   Post,
   Put,
+  Request,
   UseGuards,
   UseInterceptors,
   UsePipes,
@@ -32,11 +33,15 @@ export class UserController {
   }
   @Post()
   @UsePipes(ValidationPipe)
-  async insert(@Body() user: UserDto): Promise<UserDto> {
-    return (await this.userService.insert(user)) as
-      | UserDto
-      | { message: string }
-      | any;
+  @UseGuards(AuthGuard('jwt'))
+  async insert(@Body() user: UserDto, @Request() req) {
+    if (req.user.role === 'Admin') {
+      return (await this.userService.insert(user)) as
+        | UserDto
+        | { message: string }
+        | any;
+    }
+    return { message: 'You are not authorized to perform this action' };
   }
   @Post('login')
   async login(@Body() user: UserDto): Promise<UserDto> {
@@ -46,12 +51,19 @@ export class UserController {
       | any;
   }
   @Put(':id')
-  async update(@Body() updateUser: UserDto, @Param() params): Promise<UserDto> {
-    const oldUser = await this.userService.findById(params.id);
-    return await this.userService.update(oldUser, updateUser, params);
+  async update(@Body() updateUser: UserDto, @Param() params, @Request() req) {
+    if (req.user.role === 'Admin') {
+      const oldUser = await this.userService.findById(params.id);
+      return await this.userService.update(oldUser, updateUser, params);
+    }
+    return { message: 'You are not authorized to perform this action' };
   }
   @Delete(':id')
-  async delete(@Param() Param): Promise<UserDto> {
-    return (await this.userService.delete(Param.id)) as UserDto;
+  async delete(@Param() Param, @Request() req) {
+    if (req.user.role === 'Admin') {
+      return (await this.userService.delete(Param.id)) as UserDto;
+    }
+
+    return { message: 'You are not authorized to perform this action' };
   }
 }
